Refactor: 헤더, content 재조정
