+++ src/components/AppContent.js
@@ -12,7 +12,7 @@ const AppContent = () => {
           {routes.map((route, idx) => {
             return (
               route.element && (
-                <Route key={idx} path={route.path} exact={route.exact} name={route.name} element={<route.element />} />
+                <Route key={idx} path={route.path} exact={route.exact}  name={route.name} element={<route.element />} />
               )
             )
           })}
