+++ src/components/AppHeader.js
@@ -1,100 +1,117 @@
-// import React from 'react'
-// import styled from 'styled-components'
-// import {Link, useNavigate} from 'react-router-dom'
-// import packageJson from '../../package.json'
-// import {ReactComponent as NeordinaryLogo} from '../assets/neordinary-logo.svg'
+import { ReactComponent as Logo } from '../assets/logo.svg';
+import { ReactComponent as Search } from '../assets/search.svg';
+import { ReactComponent as More } from '../assets/more.svg';
+import { ReactComponent as Notibtn } from '../assets/notibtn.svg';
+import { Routes, Route, Link, useNavigate, useLocation, Outlet, useParams } from 'react-router-dom';
+import { useDispatch, useSelector } from 'react-redux';
+import { login, logout, updateuser } from '../store/actions/login';
+import styles from './Navbar.module.css';
 
-// const AppHeader = () => {
-//   const navigate = useNavigate()
+function AppHeader(props) {
+  let location = useLocation();
+  let auth = useSelector((state) => state.auth);
+  let dispatch = useDispatch();
 
-//   // const dispatch = useDispatch()
-//   // const sidebarShow = useSelector((state) => state.sidebarShow)
-//   // const [visible, setVisible] = useState(false)
+  return (
+    <header className={styles['App-header']}>
+      <div className={styles['mainbar']}>
+        <div className={styles['top_logo']}>
+          <button type="button" className={styles['MainBar_hamberger']}>
+            <img src="https://image.wanted.co.kr/optimize?src=https%3A%2F%2Fstatic.wanted.co.kr%2Fimages%2Ficon-menu.png&amp;w=17&amp;q=75" alt="hamberger menu" />
+          </button>
+          <div>
+            <Logo onClick={() => props.navigate('/')} />
+          </div>
+        </div>
+        <div className={styles['mainmenu']}>
+          <ul className={styles['navbar-list']}>
+            <li className={`${styles['navbar-item']} ${location.pathname === '/employment' ? styles['active'] : ''}`} onClick={() => props.navigate('/employment')}>채용 이벤트</li>
+            <li className={`${styles['navbar-item']} ${location.pathname === '/resume' ? styles['active'] : ''}`} onClick={() => props.navigate('/resume')}>이력서</li>
+            <li className={`${styles['navbar-item']} ${location.pathname === '/social' ? styles['active'] : ''}`} onClick={() => props.navigate('/social')}>소셜</li>
+            <li className={`${styles['navbar-item']} ${location.pathname === '/freelancer' ? styles['active'] : ''}`} onClick={() => props.navigate('/freelancer')}>프리랜서</li>
+            <li className={`${styles['navbar-item']} ${location.pathname === '/ai-prediction' ? styles['active'] : ''}`} onClick={() => props.navigate('/ai-prediction')}>AI 합격예측</li>
+          </ul>
+        </div>
+        <div className={styles['aside']}>
+          <ul className={styles['aside-list']}>
+            <li className={styles['aside-item']}>
+              <button type="button" className={styles['search-button']}>
+                <Search onClick={() => props.navigate('/search')} />
+              </button>
+            </li>
+            {auth.isLoggedIn &&
+              <li className={styles['aside-item']}>
+                <button type="button" className={styles['notiButton']}>
+                  <Notibtn width="18" height="18" className={styles['notibtn']}></Notibtn>
+                </button>
+              </li>}
+            {auth.isLoggedIn ?
+              <li className={styles['aside-item']}>
+                <button type="button" className={styles['user-profileButton']} onClick={() => { dispatch(logout()); }}>
+                  <div className={styles['avatarBorder']}>
+                    <div className={styles['avatarImage']}></div>
+                  </div>
+                </button>
+              </li>
+              : <li className={styles['aside-item']}>
+                <button type="button" className={styles['login-button']} onClick={() => props.navigate('/login/email')}>
+                  회원가입/로그인
+                </button>
+              </li>}
+            <li className={styles['aside-item4']}>
+              {/* <span className={styles['vertical-line']}>ㅣ</span> */}
+            </li>
+            <li className={styles['aside-item']}>
+              <button type="button" className={styles['company-service-button']} onClick={() => props.navigate('/company-service')}>
+                기업서비스
+              </button>
+            </li>
+          </ul>
+        </div>
+      </div>
+      <div className={styles['mainbar2-1']}>
+        <div className={styles['top_logo']}>
+          <button type="button" className={styles['MainBar_hamberger']}>
+            <img src="https://image.wanted.co.kr/optimize?src=https%3A%2F%2Fstatic.wanted.co.kr%2Fimages%2Ficon-menu.png&amp;w=17&amp;q=75" alt="hamberger menu" />
+          </button>
+          <div>
+            <Logo onClick={() => props.navigate('/')} />
+          </div>
+        </div>
+        {auth.isLoggedIn ?
+          <button type="button" className={styles['user-profileButton']} onClick={() => { dispatch(logout()); }}>
+            <div className={styles['avatarBorder']}>
+              <div className={styles['avatarImage']}></div>
+            </div>
+          </button>
+          : <div className={styles['login-button2-container']}>
+            <button type="button" className={styles['login-button2']} onClick={() => props.navigate('/login/email')}>
+              회원가입하기
+            </button>
+          </div>}
+      </div>
+      <div className={styles['mainbar2-2']}>
+        <div className={styles['mainmenu']}>
+          <ul className={styles['navbar-list']}>
+            <li className={`${styles['navbar-item']} ${location.pathname === '/employment' ? styles['active'] : ''}`} onClick={() => props.navigate('/employment')}>채용 이벤트</li>
+            <li className={`${styles['navbar-item']} ${location.pathname === '/resume' ? styles['active'] : ''}`} onClick={() => props.navigate('/resume')}>이력서</li>
+            <li className={`${styles['navbar-item']} ${location.pathname === '/social' ? styles['active'] : ''}`} onClick={() => props.navigate('/social')}>소셜</li>
+            <li className={`${styles['navbar-item']} ${location.pathname === '/freelancer' ? styles['active'] : ''}`} onClick={() => props.navigate('/freelancer')}>프리랜서</li>
+            <li className={`${styles['navbar-item']} ${location.pathname === '/ai-prediction' ? styles['active'] : ''}`} onClick={() => props.navigate('/ai-prediction')}>AI 합격예측</li>
+          </ul>
+        </div>
+        <div>
+          <button type="button" className={styles['search-button']}>
+            <Search onClick={() => props.navigate('/search')} />
+          </button>
+          <button type="button" className={styles['more-button']}>
+            <More />
+          </button>
+        </div>
+      </div>
+    </header>
+  );
+}
 
-//   return (
-//     <Header>
-//       <Wrap>
-//         <NeordinaryLogo width='30rem' />
-//         <HeaderText>
-//           {process.env.REACT_APP_ENV !== 'production' ? (
-//             <>
-//               REACT_APP_ENV : ${process.env.REACT_APP_ENV}
-//               <br />
-//               REACT_APP_API : ${process.env.REACT_APP_API}
-//               <br />
-//               프로젝트 : ${packageJson.name}
-//               <br />
-//               설명 : ${packageJson.description}
-//               <br />
-//               버전 : ${packageJson.version}
-//             </>
-//           ) : (
-//             ''
-//           )}
-//         </HeaderText>
-//       </Wrap>
-//       <Wrap>
-//         <Link to='/login'>
-//           <Button>로그아웃</Button>
-//         </Link>
-//         <Button
-//           onClick={() => {
-//             // replace true 는 뒤로가기시 이전 페이지로 안가짐
-//             navigate('/dashboard', {replace: true})
-//           }}
-//         >
-//           대시보드로 화면 전환
-//         </Button>
-//         <Button
-//           onClick={() => {
-//             // replace false (혹은 생략) 는 뒤로가기시 이전 페이지 가짐
-//             navigate('/')
-//           }}
-//         >
-//           홈
-//         </Button>
-//       </Wrap>
-//     </Header>
-//   )
-// }
+export default AppHeader;
 
-// const Header = styled.div`
-//   width: 100%;
-//   height: 11rem;
-//   background-color: red;
-//   display: flex;
-//   flex-direction: row;
-//   align-items: center;
-//   justify-content: space-between;
-//   box-shadow: 0 3px 16px 0 rgba(0, 0, 0, 0.06);
-// `
-
-// const Wrap = styled.div`
-//   display: flex;
-//   align-items: center;
-//   background-color: green;
-//   margin-left: 5rem;
-//   margin-right: 5rem;
-// `
-
-// const HeaderText = styled.p`
-//   font-size: 1.2rem;
-//   font-weight: 600;
-//   color: white;
-//   font-family: 'Apple SD Gothic Neo', 'Noto Sans KR', sans-serif;
-// `
-
-// const Button = styled.div`
-//   padding: 1rem;
-//   margin-left: 1.5rem;
-//   margin-right: 1.5rem;
-//   border-radius: 0.6rem;
-//   background-color: blue;
-//   font-size: 1.2rem;
-//   font-weight: 600;
-//   color: white;
-//   text-align: center;
-//   cursor: pointer;
-// `
-
-// export default AppHeader
